pyproject.toml
[project]
name = "ask-gemini"
version = "0.1.0"
description = "A simplest MCP example showing to expose Gemini + Web search as a tool"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastmcp>=2.5.2",
    "fire>=0.7.0",
    "google-genai>=1.18.0",
    "httpx>=0.28.1",
]
